Roi unloading

+++ ArmsModels/BaseModels/Operations/ROI/ROIDriverIncentiveModel.cs
@@ -17,7 +17,7 @@ namespace ArmsModels.BaseModels
         }
         public int? ID { get; set; }
         [Required]
-        public int? Wheels { get; set; }
+        public byte? Wheels { get; set; }
         [Required]
         public ROITonnageModel BSType { get; set; } = new();
         public OrderModel Order { get; set; } = new();
