+++ ArmsServices/DataServices/Operations/ROI/ROIDriverBattaService.cs
@@ -69,7 +69,7 @@ namespace ArmsServices.DataServices
             return new ROIDriverBattaModel
             {
                 ID = dr.GetInt32("ID"),
-                Wheels = dr.GetInt32("Wheels"),
+                Wheels = dr.GetByte("Wheels"),
                 BodyType = dr.GetString("BodyType"),
                 FromStdKM = dr.GetDecimal("FromStdKM"),
                 ToStdKM = dr.GetDecimal("ToStdKM"),
